devcpu-data: Provide a file for loading at boot time
Extend the port so that users can easily apply microcode updates using
the method implemented in src r337715.  For now, this just consists of
providing all microcode update files concatenated together as a single
blob in /boot/firmware.  I also updated the package message to describe
the new method, and would appreciate any feedback on that.

Once the port update is committed and available via pkg, I'll make an
announcement on -current with more details.