net/libslirp: Make purecap libslirp.so usable
The UDP implementation input path does a gross thing: it overwrites the
first 8 bytes of the IP header, which it doesn't care about after IP
processing, with a backpointer to the mbuf.  When the pointer size isn't
4 or 8, this of course doesn't work and the end result is that
udp_input() discards all packets due to checksum errors.

Happily, it doesn't make use of this backpointer, so we can simply
remove it.

The TCP case is a bit more complicated: it actually makes use of the
backpointer (during fragment reassembly) so we can't just remove it.
However, it is stored in a different pseudo-header structure which isn't
required to overlay the IP+TCP headers; it stores both v4 and v6
addresses.  The input path makes sure to leave enough headroom in input
mbufs to provide space for the pseudo-header during protocol processing,
so the extra size constraint of keeping a backpointer doesn't hurt.

I am not convinced that the current code always ensures that the
pseudo-header is sufficiently aligned; however, in practice it is for v4
traffic with no IP options, which is good enough for bhyve's slirp
backend, which is what motivates this patch in the first place.  I don't
have a convenient way to test v6 and there is no test suite.  If
insufficient alignment means that the backpointer's tag bit is cleared,
at least we'll fail closed.