security/sssd2: Use kvm_getprocs(3) to get the process list
This is basically equivalent to using sysctl() directly, but requires
less code, and kvm_getprocs() has internal retries to handle the race
condition where the process list grows after fetching the size and
before fetching the list.

The previous code could fail if the process list grows very quickly
while fetching the list, so this approach is preferable.